Prevent "dependency confusion" resulting in a malicious upstream gem taking precedence over a local gem
Rather than combine all gems from BOTH repositories, only include
upstream gems if a gem with the same name is NOT PRESENT LOCALLY.

This is a breaking behavior change in the interest of shipping Geminabox
in a "secure by default" configuration.

To restore the previous behavior, exposing yourself to this vuln:
* Set env var: `RUBYGEMS_PROXY_MERGE_STRATEGY=combine_local_and_remote_gem_versions`
* Add to config.ru: `Geminabox.rubygems_proxy_merge_strategy = :combine_local_and_remote_gem_versions`

One reason to keep the old behavior is if you use Geminabox to override
a public gem version with your own private gem build. In this case,
recommend using Bundler to pin to a private gem source or releasing a
forked gem with a different version instead.